Add CDN for the SpotifyWebApi library and minor cleanup
Creates a getToken method to get an API token, because there's no
getToken method on the SpotifyWebApi.

Attempts to tweak the searchTrack response, I started lookign at href
instead of preview_url. The preview_url was null for me, so I'm not sure
what I'm doing wrong. Also I'm seing a 401 when I try to make a request
to get the track aduio with the href provided.